Update to url 1.0.0
rust-url v1.0.0 had some major breaking changes.
The biggest one was the change from struct members to getter methods
and the fact that the scheme mapper was removed.
With it, the `InvalidScheme` error was removed.

redis-rs used this to fail with schemes other than `redis` and `unix`.

I now changed the return type to a RedisResult, so we can handle the
error ourselves. It is only called at one position anyway.

`parse_redis_url` _is_ part of the publicly exposed API though, so
changing the return type is a breaking change.