Relax ScriptInvocation lifetimes
The `arg` and `key` methods could only be used in a chain as
demonstrated in the docs. If `arg` and `key` are to be called
programmatically as in

    for arg in &args {
        invocation.arg(arg);
    }

There was a borrowing error, and `invoke` could never be called. This
detaches the lifetime of the mutable borrow from the lifetime of
ScriptInvocation (specifically, says it will be shorter) to support the
above use case.