Fix PubSub cleanup
When a subscribe or unsubscribe command is sent to the server, the
connection enters a state where the only valid commands are pubsub
commands. To return to the normal state, all subscriptions must be
dropped. This is achieved by sending an unsubscribe and punsubscribe
command to the server. The server then replies with all of the
subscriptions that were dropped.

This patch ensures that all messages are received from the server after
sending those commands and that the connection object is again usable
for non-pubsub activities.

The PubSub API was changed to reflect the fallible nature of this
operation. In the event of an error during unsubscribing, the connection
is simply lost. In practical terms, this means that entering the PubSub
state consumes the connection, and the connection is retrieved through
calling `into_inner`. This will likely cause issues with connection
pooling libraries since they don't provide ownership.