Alternative PubSub cleanup model
This implementation adds a flag to the `Connection` struct to indicate
whether the connection is stuck in the PubSub state after PubSub is
dropped. Subsequent commands check this flag and reattempt the
unsubscribe routine to put the connection back into a known good state.

The main drawback to this implementation is the additional state inside
`Connection`.