Expose connection status in `is_open` method. (#150)
* Expose connection status in `is_open` method.

The `is_open` method can be used in `r2d2::ManageConnection#has_broken`
to determine if the connection was closed or failed for an other reason.

Before `r2d2-redis` had no way to check if the connection broke during
the time a request was made. This meant the only way to handle
connection errors (either problem with the network or redis server) was
to use `test_on_check_out(true)` which run the `is_valid` check before
a connection is checked out of the pool.

While this works it has an impact on the throughput because every
checkout makes a request to redis to check the connection status.

Exposing the connection status allows the user to optimize for the
"golden path" and only mark connections (and thus removing them from the
pool) if something went wrong.

As far as I know the only way to check the status of a connection is to
either read or write to it, writing arbitrary data would mess with a
working connections and trying to read data could block. That is why I
opted to only update the status when parsing a response failed with
a `ResponseError`. Not sure why but this also works when the server
closed the connection while writing to a TCP socket (verified manually
with toxiproxy).

* Make ActualConnection enum variants hold a Tcp/UnixConnection struct

Allows to deconstruct the struct and only use the fields that are
required.

* Incorporate feedback