fix(async): Don't block the executor from shutting down
The task that drives `SharedConnection` is supposed to shutdown
automatically once the last reference to it is dropped. It did not do
that due to the `poll_read` call in `close` which registered another
`read` on the socket. Since this socket only produces data in response
to any requests done to redis and the last `SharedConnection` reference
were dropped there will only be more data to read while there are
requests in flight.

This ensures that we only drive the socket until those requests are done
whereafter we shut it down.

Tests did not catch this because they only waited for the main task to
finish which were fixed with `block_on_all`