fix(script): invoke_async returning complex type
Previously, invoking a script with a complex return type would cause the following error: Response was of incompatible type: "Not a bulk response" (response was string-data('"4b98bef92b171357ddc437b395c7c1a5145ca2bd"'))
This was because the Future returned when loading the script into the DB returns the hash of the script, and thus the return type of String would not match the intended return type.
This adds an enum to account for the different Future return types.