feat: Add Cmd::set, Cmd::get etc
It is currently possible to call `set`,`get` etc on a connection and
run the query directly but it isn't possible to create a `Cmd` struct
without running the query (except for `Pipeline`).

This adds this capability with just some small tweaks to the macro
allowing `Cmd::set("abc", 123)` to be used etc.