Implement non-parallel reconnections using shared future
This way, the first task that notices a connection loss will initiate a
reconnection, while the other tasks wait for the connection future to
complete.

In order to avoid having to lock a mutex for replacing the future,
the arc-swap crate is used. It performs well in read-often update-seldom
scenarios.