Fix incorrect routing for XREAD/XREADGROUP command
The XREAD/XREADGROUP commands can have a variable number of args,
and proper routing of the command relies on finding the key after
the "STREAMS" keyword. This was failing to properly route the if
the keyword was not uppercase. Fixes the matching logic make each
arg uppercase before comparing to the "STREAMS" keyword.