Support for requesting `Cmds` directly from a `Connection`
Adds a new function to the `ConnectionLike` trait that accepts a
reference to a `Cmd`. This allows different connections to handle
the `Cmd` as necessary. For a single Redis instance, there is no
difference, but for Redis Cluster it allows the `Cmd` to be routed
without needing to first parse it back into a `Value`.

This results in a *slight* performance increase when issuing
commands with a `ClusterConnection`, with the performance benefits
becoming more prominent with more commands. With the upcoming work
to add pipeline support with Redis Cluster (#307), I observed a
42% reduction in execution time when issuing a pipeline query with
100 elements.

For backwards compatability, a new trait `Routable` is introduced
and the code that determines which node(s) should handle a
particular command is made generic over said trait.