Add support for `ZRANGESTORE`.
Switch to a single zrangestore cmd receiving an options enum.

There's still plenty to dislike about this approach compared to the
previous attempt where the combinations of various optional args meant
we had an explosion of functions. That said, this does seem a bit more
straightforward to read and write.

This switch diverges from the design of the earlier sorted set commands
(which the previous attempt strived to mimic), so we might need to loop
back around at a later date with some breaking changes to the earlier
APIs to help steer closer to consistency.

Switch to 3 variants of zrangestore w/ owned builder for options.

Splitting the difference seems to make some sense to simplify the logic
required in the `ZRangeStoreOptions` type.

By moving the management of the exclusive options (byscore, bylex) to command
functions instead of the options type, we can closer mimic the other options
types already defined in the crate.

Add ZRangeStoreOrdering enum, remove order-based variants funcs.

Instead of relying on the ordering being specified via the command
function, it's now tracked as a field of the ZRangeStoreOptions struct.

whitespace

Add tests for zrangestore.

better examples