perf: Don't enqueue multiplexed commands if the receiver is dropped
It is already ambiguous whether the command will run or not if the
receiver has dropped (it could have been dropped before the message was
sent over the channel). So if we see that there is no one to receive the
response before we have had a chance to send it over the wire then we
immediately drop the message instead.

This can help reduce load on the connection in circumstances where the
caller is overloaded and starts timing out on the requests.