Improve MultiplexedConnection Error Handling
It was recently discovered that `aio::MultiplexedConnection`
does not properly handle errors. If a command in the pipe fails,
responses from subsequent commands in the pipe are not processed
and will be returned to the next request issued on the connection,
with potentially catastrophic results. This fix ensures that
all pending commands in the pipe are processed before returning.

Fixes #698.