Enable discovery of aborted Connections
This will allow connection managers like bb8 to drop async Connections
when the Future that owned them was dropped.
See: https://github.com/djc/bb8/blob/main/redis/src/lib.rs#L78-L80

Futures drop synchronously, so there's no practical way to block them
until their response from Redis has been fully hydrated. Instead,
this change _only_ seeks to make such a situation knowable to a
ConnectionManager so it can make a choice as to what to do with the
connection.

The impact of reusing a connection in this state has no upper bound;
it violates privacy of unrelated values. This change addresses this
by adding an IO error in the Connection under ConnectionLike such
that if the connection is in some partial & unknown state it will
refuse to execute the requested commands.

By using IoError as the ErrorKind, it is believed that other pools
will be able to make a good choice about what to do when they find
a Connection in this state, no matter how it got there.

Relates to: https://github.com/redis-rs/redis-rs/issues/325