Common slot parsing
Use common `parse_slots` command; this allows for eliminating
a lot of largely duplicate code in the cluster-async module but
requires adding support for `ClusterParams` to track connection
settings for TlsMode, username, password, etc., which was previously
(incompletely) tracked by the connection strings constituting the keys
of the ConnectionMap.

These changes also have the effect of enabling support in the cluster-
async module for insecure TLS and both usernames and passwords for
authentication.