Add API to get whole batches from cursors
This solves a limitation where aggregating the results of a cursor into
a vector required a O(n) move between the Vec owned by redis-rs and a
Vec owned by the application.

The new API is exposed similarly to the existing async_iter() method on
`Cmd` with the suffix `_batched`. The new batched cursor is used
internally by the orignal AsyncIter API to reduce code duplication and
also increase test coverage of the new iterator type.