Fix pipeline routing.
This fixes 2 issues with the previous implementation of `route_pipeline`
1. if the first command didn't have specific slot routing then
`iter.next().map(route_for_command)?;` caused the function to return
`None`, even if later commands had specific routes.
2. If 2 commands had conflicting routes then the function would return
`None`, instead of choosing one of the routes.

A `None` return value meant that a connection would be chosen randomly.
This change ensures that if there is at least a partially valid route
 to be chosen, the client will choose it.