Allow configuration of cluster retry wait duration (#859)
* Allow configuration of cluster retry wait duration

This change allows the user to configure the wait duration between retry
attempts in the cluster clients.

This change also introduces jitter, so that if multiple requests
are waiting, they won't be all sent at exactly the same time.

* Test cluster: wait for status ok in log file.

During the development of the feature I ran tests with lower wait time
between retries, and often received CLUSTERDOWN errors. This means
that RedisCluster wasn't ready to use when constructed.