Improve cluster retries
Currently we don't sleep for certain classes of errors, which in my
testing leads to certain potentially suboptimal results. Most notably,
`MOVED` errors are likely to be thrown when a cluster is in an
unhealthy state, such as during a node loss. With our current strategy
of immediately retrying `MOVED` errors without pause, we eliminate
the client's ability to recover in many scenarios.

In any case, I think it's reasonable that we sleep for all retryable
errors.

Relates to #906.