When executing:   async_conn.hscan_match::<_, _, Vec<u8>>(name, prefix.as_slice()).await? or   cmd("HSCAN")         .arg(name).cursor_arg(0).arg("match").arg(prefix).clone()         .iter_async::<Vec<u8>>(&mut async_conn)         .await?;
When no data with prefix is found, it will be blocked.

Viewing the source code is due to the blockage caused by the cursor not being 0 but the return result set being empty, which will always attempt to obtain more data.

The following is the result of directly executing the Redis command in Redis cli:
127.0.0.1:6379> hscan "map001" 0 match "10206*" count 10
1) "1572864"
2) (empty array)