Remove allocation from `command` function. (#1008)
Instead of allocating the ASCII uppercase conversion of
`self.arg_idx(1)` and then copying it into the target vec, we copy the
slice, and then convert it to uppercase in-place. This saves the
allocation of the vec, which is only used for copying.

The perf difference here is too small to be conclusive in benchmarks,
but the reduction in the generated assembly is noticable
(~920 lines of assembly to ~800).