Handle server errors in array response 2.
Currently server errors stop the parser and return a RedisError. This caused errors that are returned inside an array, such as transaction errors, to cause the rest of the array to not be parsed.
This is solved by adding a `ServerError` variant to `Value` that contains the server errors, so when parsing to the server message, the array will finish parsing, and then extracting the error.