Merge push manager change from redis-rs. (#125)
* Initial implementation of RESP3 (#757)

These changes implement all RESP3 types (excluding streamed types). 
RESP3 can be enabled per connection by adding `?resp3=true` to 
connection uri. It currently supports PubSub as RESP2 PubSub support 
in library, but in future PRs it'll support handling normal commands 
and PubSub messages in one connection. Only `num-bigint` is added 
as dependency to support `BigNumber` type.

Changes made in support of #329 and #749

* Add RESP3 support to cluster connections. (#1001)

* Resp3 Push Management (#898)

* squash!

* oops!

* test invalidation in cluster && introduce client tracking options

* introduce basic PubSub functionality to MultiplexedConnection and make tokio sender unbounded channel

* fix tests & linter, make PushSender::Tokio as aio feature only

* add resp3 to ci branches and fix cluster client tracking option

* test dropping and update ci yml

* remove unsubscribe fn and introduce unsubscribing by dropping receiver.

* fix tests because RedisJson returns responses in an array when it's resp3

* override redisjson cache (it's a temp solution)

* add -skip test_module to RESP3 testing and upgrade redis 6.2.4 to 6.2.13

* test json modules with RESP3 and get json fix from main

* in redis v7 RedisJson is different with Resp3 :thinking:

* Implement Pub/Sub in Async Cluster & fmt & remove usage of is_err_and(stable only after v1.70)

* don't use sharded pub/sub with redis v6

* use REDIS_VERSION in env instead of using HELLO command to fetch redis version

* oops

* fix linter

* fix fmt and remove benchmark from CI

* simplify PushManager and add tokio as non-optional dependency.

* get fixes from 220c6a9742ce4dd855f02f471f4cb03469b3610f

* use --test-threads=1

* override redisjson cache (it's a temp solution)

* remove get_push_manager from traits & remove push manager from aio::Connection

* remove client_tracking_options

* remove 0.21.x from rust.yml

* add tests for pushmanager

* format & move push_info into a variable

* change tests according to comments.

* apply 6.2.4 changes && fmt

* try to fix

* remove con_addr & remove pub/sub support in cluster connections

* add disconnection handling to sync, mpx, cm && test it

* remove push_manager argument from connection creation

* better docs

* add has_reply function to PushKind

* remove no response command support in mpx since it's not used in mpx pub/sub

* apply changes from #994

* fix tests

* Use enum instead of boolean for RESP version. (#1012)

Since there's a discussion starting about what might become RESP4, this
PR will make it easier to add more RESP versions in the future.

* Add lock file to keep MSRV constant. (#1039)

* [opt] preallocate buffer for evalsha in `Script` (#1044)

* Cluster: fix read from replica & missing slots. (#965)

* Cluster SlotMap: Allow missing slots.

The cluster can still operate when some slots don't have coverage, so
there's no reason that the connection won't work.
Requests that should be routed to missing slots will be forwarded to a
random node.

* Fix SlotAddrs not holding all replicas.

This change fixes 2 issues -
1. Since SlotAddrs saves at most 1 replica (or none, if
`read_from_replica` is false), it means that when there are 2 or more
replicas in the cluster, `AllNodes` commands aren't actually sent to
all nodes.
2. `SlotAddr::Replica` carries 2 separate semantic meanings - it could
mean that the command is a readonly command, or that the user asked to
route the command to a replica.
The first case should be affected by the `read_from_replica` client
flag, but the second case shouldn't - if the user requested that this
specific command be routed to the replica, the specific request should
override the prior configuration.

* Fix multi-slot routes/addresses mismatch.

If the `SlotMap` is partial, and some slots don't have addresses, then
we need to communicate this back when getting addresses for multi-slot
commands. Otherwise we the key indices won't match the addresses,
and commands with the wrong keys will be sent to the wrong nodes.

* fix typo

* fix #1045: impl `Clone`/`Copy` for `SetOptions` (#1046)

* docs: add "connection-manager" cfg attr (#1048)

* Fix new clippy lints. (#1052)

Replace Vec reference with a slice, and remove iterator trait when using ExactSizeIterator, since it encompasses it.

* Rename Value enum types in order to match Redis RESP names. (#779)

* Rename Value::Bulk to Value::Array.

* Rename Value::Status to Value::SimpleString.

* Rename Value::Data to Value::BulkString.

* Fix debug names of values.

* Fix nightly lints.

---------

Co-authored-by: Altan Özlü <5479094+altanozlu@users.noreply.github.com>
Co-authored-by: Huxley Hu <framlog@users.noreply.github.com>
Co-authored-by: Ahmad <39441506+ahmadbky@users.noreply.github.com>
Co-authored-by: ⭐️NINIKA⭐️ <DCNick3@users.noreply.github.com>