Fix nightly compilation warnings.
Adds missing generic declarations or return types for `()`, when until the 2024 edition ! was inferred. This fixes this warning:

warning: this function depends on never type fallback being `()`
 --> redis/examples/async-await.rs:4:7
  |
4 | async fn main() -> redis::RedisResult<()> {
  |       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  |
  = warning: this was previously accepted by the compiler but is being phased out; it will become a hard error in a future release!
  = note: for more information, see issue #123748 <https://github.com/rust-lang/rust/issues/123748>
  = help: specify the types explicitly
note: in edition 2024, the requirement `!: FromRedisValue` will fail
 --> redis/examples/async-await.rs:8:9
  |
8 |     con.set("key1", b"foo").await?;
  |         ^^^
  = note: `-D dependency-on-unit-never-type-fallback` implied by `-D warnings`
  = help: to override `-D warnings` add `#[allow(dependency_on_unit_never_type_fallback)]`

In order to make this easier, this also changes the generics used in `Cmd/Pipeline::query_async`, so that the function declaration would only need to define a single generic value.