Remove tokio dependency from non-aio build.
This is done by changing the sync Connection's push sender to the std sender instead of tokio sender.
The decision to use the tokio sender in the sync implementation came from the previous design, where all connections used the same PushManager, and so had to use the same channel type.

https://github.com/redis-rs/redis-rs/pull/898#discussion_r1297473079