refactor redis-rs to be sans-io backwards compat (sync only)
In function of #1481,
for now as discussed the sync (non-aio) API's only,
once we get this merged I can tackle the async version as well.

I think this PR covers the entire sync surface layer,
it was a bit tricky to do this fully backwards compatible,
which also means I had to make some design choices that I
otherwise would not have done.

Do let me know if I missed something,
or need to do stuff different.
It's the first contribution to this codebase,
so I might also have done some things wrong
that are obvious to regular contributors. Do let me know.