Unify two definitions of TlsConnParams.
Previously these were defined differently, in two different places, depending on whether
`cfg(feature = "tls-rustls")` was enabled. This commit doesn't change anything except how that is
spelled in the code. Now they share a definition, but the fields still depend on the feature, same
as before.

The point is to make it possible to add features to `TlsConnParams` that are always there, or which
perhaps depend on a different feature being present.