Run most cluster tests in secure mode when rustls is enabled (#1534)
* fixup: avoid warning about unused import (fatal in `make test`)

* Run most cluster tests in secure mode when `rustls` is enabled

In trying to add a test for a new security mode in #1529, I found out no existing cluster tests run
in `TlsMode::Secure` mode. It seems like a good idea to get some coverage for that case first.

All but 6 tests pass with certificate checking enabled, so this PR transitions tests to run in
`Secure` mode by default when `rustls` is used, and explicitly opts out those 6 tests.

This is `tls-rustls` only for now. TLS support in the cluster tests works by generating a one-off
root cert, and it appears `redis-rs` doesn't support configuring a root cert in `tls-native-tls`
builds.