Implement `danger_accept_invalid_hostnames` option. (#1529)
* Unify two definitions of TlsConnParams.

Previously these were defined differently, in two different places, depending on whether
`cfg(feature = "tls-rustls")` was enabled. This commit doesn't change anything except how that is
spelled in the code. Now they share a definition, but the fields still depend on the feature, same
as before.

The point is to make it possible to add features to `TlsConnParams` that are always there, or which
perhaps depend on a different feature being present.

* Support a `danger_accept_invalid_hostnames` option.

The option is enabled when building with either `tls-native-tls` or `tls-rustls`, but it only has
an effect when using `tls-native-tls`.

* Implement accept_invalid_hostnames for rustls.

* Fix compiler error when building with `tls-rustls` only.

The field being marked with a `#[cfg]` attribute here is not present in that configuration.

* Rename to `ConnectionAddr::set_danger_accept_invalid_hostnames`.

Also, add a copy of the comment here.

* Add a test for the new option with rustls.

* Fix `cfg` attributes on `danger_accept_invalid_hostnames`

* Snapshot of debugging and trying to get test coverage stood up.

* Fix config problems with danger_accept_invalid_hostnames tests

* Factor out connection smoke-test code.

* Add async tests.

* Revert changes that turned out to be unnecessary

* Document a really weird cfg

* Remove redundant import caught by Rust 1.75.0 only (!)

* Use async feature names in async code, per a review comment.

* Fix a too-informal comment.

* Remove unneeded Default impl on this security-related type.

* Better comment strange combination of features.

There's a combination of features that causes `danger_accept_invalid_hostnames` to exist, but where
it should be an error to actually try to use it. This tries to clarify the code that implements
this weird case.

* Finish removing unwanted Default impl.