Fix unreachable error when parsing a nested tuple.
The unreachable code assumed that a nested array has to be able to converted to the target tuple, and so panicked when when the array is larger. Instead, we should just fallback to the next piece of code, and allow the array to be split into multiple returns of the target tuple.