Add Client Side Caching for async Cluster Connections (#1642)
This is quite a small change because most of the logic is already implemented in other modules used by ClusterConnection.

I wanted to consolidate cache tests nicely across mpx,cm and cluster connections but so far failed to do so.

With that reason patch adds three new tests.

I have made a behaviour change to caching, it won't cache server errors.
This was due to caching MOVED errors and invalidating after topology change,
it wouldn't cause a problem in this case but I don't think it's safe.