feat: Make transaction status of pipelines public (#1663)
When trying to mock Redis connections, the `ConnectionLike`'s
`req_packed_commands` expects a different response shape on
`transactional` and non-`transactional` pipelines.

(Vec with single entry of Redis array of results, resp. Vec of results)

In order to hide this difference from users, it would be helpful to
check if a pipeline is in transaction mode or not.

To allow this, we make the accessor function public.

Co-authored-by: Christian Aistleitner <christian@quelltextlich.at>