Fix handling of nested tuples. (#1675)
* Fix handling of nested tuples.

The blocks that tried to convert internal collections to tuples caused bad results if some of the values in the items array could be converted to tuples and others couldn't.
```rs
for item in items {
    match item {
        Value::Array(ch) => {
            if  let [$($name),*] = &ch[..] {
            rv.push(($(from_redis_value(&$name)?),*),)
            };
        },
        _ => {},

    }
}
if !rv.is_empty(){
    return Ok(rv);
}
```
So it was replaced with a block that first verifies that all of the values in the collection can be converted to the tuple.
```rs
if items.iter().all(|item|item.is_collection_of_len(n)) {
    return items.iter().map(|item|from_redis_value(item)).collect();
}
```

Fixes https://github.com/redis-rs/redis-rs/issues/1674

* further shorten the code